requirements.txt
Django==5.0.6
pandas==2.2.2
pysmb==1.2.9.1
python-dotenv==1.0.1
gunicorn==22.0.0
